IVYPORTAL-9788 Various or urgent requests in IVYPORTAL S187  - [BUG] - Clean up obsoleted data deletes wrong UserFavorites process(handle feedback)

+++ AxonIvyPortal/PortalKit/src/ch/ivy/addon/portalkit/service/CleanUpObsoletedUserDataService.java
@@ -35,6 +35,7 @@ public class CleanUpObsoletedUserDataService {
   private List<UserDTO> currentUsers;
   private List<Long> userIds;
   private Long applicationId;
+  private boolean isError;
 
   public void cleanUpData() {
     StopWatch stopWatch = new StopWatch();
@@ -69,13 +70,18 @@ public class CleanUpObsoletedUserDataService {
     UserProcessService userProcessService = new UserProcessService();
     List<UserProcess> userProcesses = userProcessService.findAll();
 
-    List<String> allUserName = new ArrayList<>();
+    List<String> userNameOnAllApps = new ArrayList<>();
     if (CollectionUtils.isNotEmpty(userProcesses)) {
-      allUserName.addAll(collectAllUserOnServer());
+      userNameOnAllApps.addAll(collectAllUserOnServer());
+    }
+    // In case we got any errors during Collect data phase
+    // Then skip clean up User Favorites data
+    if (isError) {
+      return;
     }
 
     CollectionUtils.emptyIfNull(userProcesses).stream()
-        .filter(userProcess -> StringUtils.isBlank(userProcess.getUserName()) || (!allUserName.contains(userProcess.getUserName())))
+        .filter(userProcess -> StringUtils.isBlank(userProcess.getUserName()) || (!userNameOnAllApps.contains(userProcess.getUserName())))
         .forEach(userProcess -> {
           Ivy.log().info("CLEAN_UP_JOB: Delete UserFavourite {0} of user {1}", userProcess.getProcessName(), userProcess.getUserName());
           userProcessService.delete(userProcess);
@@ -84,28 +90,40 @@ public class CleanUpObsoletedUserDataService {
 
   private List<String> collectAllUserOnServer() {
     Ivy.log().info("CLEAN_UP_JOB: Started collecting users overall apps");
-    List<String> allUsers = new ArrayList<>();
+    List<String> userNameOnAllApps = new ArrayList<>();
     List<IApplication> allApplications = collectPortalAppOnServer();
-    allApplications.forEach(app -> {
-      allUsers.addAll(findUsersByApp(app));
+    CollectionUtils.emptyIfNull(allApplications).forEach(app -> {
+      userNameOnAllApps.addAll(findUsersByApp(app));
     });
     Ivy.log().info("CLEAN_UP_JOB: Finished collecting users overall apps");
-    return allUsers;
+    return userNameOnAllApps;
   }
 
   private List<String> findUsersByApp(IApplication app) {
-    return IvyExecutor.executeAsSystem(() -> {
-      return app.getSecurityContext().getUsers().stream().map(IUser::getName).collect(Collectors.toList());
-    });
+    try {
+      return IvyExecutor.executeAsSystem(() -> {
+        return app.getSecurityContext().getUsers().stream().map(IUser::getName).collect(Collectors.toList());
+      });
+    } catch (Exception e) {
+      Ivy.log().error("CLEAN_UP_JOB: cleanUpUserFavouriteProcess - Cannot get users data", e);
+      isError = true;
+    }
+    return new ArrayList<>();
   }
 
   private List<IApplication> collectPortalAppOnServer() {
-    return IvyExecutor.executeAsSystem(() -> {
-      List<IApplication> applications =
-          ServerFactory.getServer().getApplicationConfigurationManager().getApplicationsSortedByName(false);
-      return applications.stream().filter(app -> app.findReleasedLibrary(PortalLibrary.PORTAL_STYLE.getValue()) != null)
-          .collect(Collectors.toList());
-    });
+    try {
+      return IvyExecutor.executeAsSystem(() -> {
+        List<IApplication> applications =
+            ServerFactory.getServer().getApplicationConfigurationManager().getApplicationsSortedByName(false);
+        return applications.stream().filter(app -> app.findReleasedLibrary(PortalLibrary.PORTAL_STYLE.getValue()) != null)
+            .collect(Collectors.toList());
+      });
+    } catch (Exception e) {
+      Ivy.log().error("CLEAN_UP_JOB: cleanUpUserFavouriteProcess - Cannot get application info", e);
+      isError = true;
+    }
+    return new ArrayList<>();
   }
 
   private void cleanUpUserTaskFilter() {
